completed all todo's

+++ CS321_W5D2_BlogAPI.Core/Services/PostService.cs
@@ -19,10 +19,15 @@ namespace CS321_W5D2_BlogAPI.Core.Services
 
         public Post Add(Post newPost)
         {
-            // TODO: Prevent users from adding to a blog that isn't theirs
-            //     Use the _userService to get the current users id.
-            //     You may have to retrieve the blog in order to check user id
-            // TODO: assign the current date to DatePublished
+            var currentUserId = _userService.CurrentUserId;
+
+            var blog = _blogRepository.Get(newPost.BlogId);
+
+            if (currentUserId != blog.UserId)
+            {
+                throw new ApplicationException("You can not post to a blog that does not belong to you.");
+            }
+            newPost.DatePublished = DateTime.Now;
             return _postRepository.Add(newPost);
         }
 
@@ -44,15 +49,28 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         public void Remove(int id)
         {
             var post = this.Get(id);
-            // TODO: prevent user from deleting from a blog that isn't theirs
+            var currentUserId = _userService.CurrentUserId;
+            var blog = _blogRepository.Get(post.BlogId);
+
+            if (currentUserId != blog.UserId) // != post.Blog.UserId
+            {
+                throw new ApplicationException("You can not delete a post that does not belong to you.");
+            }
             _postRepository.Remove(id);
         }
 
         public Post Update(Post updatedPost)
         {
-            // TODO: prevent user from updating a blog that isn't theirs
+            var currentUserId = _userService.CurrentUserId;
+
+            var blog = _blogRepository.Get(updatedPost.BlogId);
+
+            if (currentUserId != blog.UserId)
+            {
+                return null;
+            }
+
             return _postRepository.Update(updatedPost);
         }
-
     }
 }
